refactor(routes): clarify event and moment navigation
Move the events and moments detail flows into route-owned stacks so those screens are managed by the app-level navigator instead of living inside the tab navigator.

Keep the tab folders as Slot layouts so the tabs still resolve to the clean events and moments route names without creating another nested stack.

Switch auth handling in the root layout to Stack.Protected so access control lives in the router configuration instead of custom redirect checks.
